+++ modules/AWSPowerShell/Cmdlets/EC2/Basic/Edit-EC2VpcEndpoint-Cmdlet.cs
@@ -123,6 +123,19 @@ namespace Amazon.PowerShell.Cmdlets.EC2
         public System.Boolean? PrivateDnsEnabled { get; set; }
         #endregion
         
+        #region Parameter DnsOptions_PrivateDnsOnlyForInboundResolverEndpoint
+        /// <summary>
+        /// <para>
+        /// <para>Indicates whether to enable private DNS only for inbound endpoints. This option is
+        /// available only for services that support both gateway and interface endpoints. It
+        /// routes traffic that originates from the VPC to the gateway endpoint and traffic that
+        /// originates from on-premises to the interface endpoint.</para>
+        /// </para>
+        /// </summary>
+        [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]
+        public System.Boolean? DnsOptions_PrivateDnsOnlyForInboundResolverEndpoint { get; set; }
+        #endregion
+        
         #region Parameter RemoveRouteTableId
         /// <summary>
         /// <para>
@@ -259,6 +272,7 @@ namespace Amazon.PowerShell.Cmdlets.EC2
                 context.AddSubnetId = new List<System.String>(this.AddSubnetId);
             }
             context.DnsOptions_DnsRecordIpType = this.DnsOptions_DnsRecordIpType;
+            context.DnsOptions_PrivateDnsOnlyForInboundResolverEndpoint = this.DnsOptions_PrivateDnsOnlyForInboundResolverEndpoint;
             context.IpAddressType = this.IpAddressType;
             context.PolicyDocument = this.PolicyDocument;
             context.PrivateDnsEnabled = this.PrivateDnsEnabled;
@@ -322,6 +336,16 @@ namespace Amazon.PowerShell.Cmdlets.EC2
             if (requestDnsOptions_dnsOptions_DnsRecordIpType != null)
             {
                 request.DnsOptions.DnsRecordIpType = requestDnsOptions_dnsOptions_DnsRecordIpType;
+                requestDnsOptionsIsNull = false;
+            }
+            System.Boolean? requestDnsOptions_dnsOptions_PrivateDnsOnlyForInboundResolverEndpoint = null;
+            if (cmdletContext.DnsOptions_PrivateDnsOnlyForInboundResolverEndpoint != null)
+            {
+                requestDnsOptions_dnsOptions_PrivateDnsOnlyForInboundResolverEndpoint = cmdletContext.DnsOptions_PrivateDnsOnlyForInboundResolverEndpoint.Value;
+            }
+            if (requestDnsOptions_dnsOptions_PrivateDnsOnlyForInboundResolverEndpoint != null)
+            {
+                request.DnsOptions.PrivateDnsOnlyForInboundResolverEndpoint = requestDnsOptions_dnsOptions_PrivateDnsOnlyForInboundResolverEndpoint.Value;
                 requestDnsOptionsIsNull = false;
             }
              // determine if request.DnsOptions should be set to null
@@ -426,6 +450,7 @@ namespace Amazon.PowerShell.Cmdlets.EC2
             public List<System.String> AddSecurityGroupId { get; set; }
             public List<System.String> AddSubnetId { get; set; }
             public Amazon.EC2.DnsRecordIpType DnsOptions_DnsRecordIpType { get; set; }
+            public System.Boolean? DnsOptions_PrivateDnsOnlyForInboundResolverEndpoint { get; set; }
             public Amazon.EC2.IpAddressType IpAddressType { get; set; }
             public System.String PolicyDocument { get; set; }
             public System.Boolean? PrivateDnsEnabled { get; set; }