+++ modules/AWSPowerShell/Cmdlets/IdentityManagement/Basic/Get-IAMOrganizationsAccessReport-Cmdlet.cs
@@ -42,7 +42,7 @@ namespace Amazon.PowerShell.Cmdlets.IAM
     /// this operation. For more information, see <a href="https://docs.aws.amazon.com/IAM/latest/UserGuide/access_policies_access-advisor.html">Refining
     /// permissions using service last accessed data</a> in the <i>IAM User Guide</i>.
     /// </para><para>
-    /// For each service that principals in an account (root users, IAM users, or IAM roles)
+    /// For each service that principals in an account (root user, IAM users, or IAM roles)
     /// could access using SCPs, the operation returns details about the most recent access
     /// attempt. If there was no attempt, the service is listed without details about the
     /// most recent attempt to access the service. If the operation fails, it returns the