+++ modules/AWSPowerShell/Cmdlets/IdentityManagement/Basic/New-IAMRole-Cmdlet.cs
@@ -135,7 +135,9 @@ namespace Amazon.PowerShell.Cmdlets.IAM
         /// <para>
         /// <para>The name of the role to create.</para><para>IAM user, group, role, and policy names must be unique within the account. Names are
         /// not distinguished by case. For example, you cannot create resources named both "MyResource"
-        /// and "myresource".</para>
+        /// and "myresource".</para><para>This parameter allows (through its <a href="http://wikipedia.org/wiki/regex">regex
+        /// pattern</a>) a string of characters consisting of upper and lowercase alphanumeric
+        /// characters with no spaces. You can also include any of the following characters: _+=,.@-</para>
         /// </para>
         /// </summary>
         #if !MODULAR